LIBCIR-290. Modified "Dockerfile.prod" to support "PM2" process manager
Modified the "Dockerfile.prod" used for the production Kubernetes Docker
image to add the "pm2" process manager, and use "pm2" to run Angular.

Added "dspace-ui.json" file to specify the configuration parameters for
"pm2".

The "--no-daemon" flag is used to start pm2, as otherwise the Docker
container simply exits after starting pm2.

Explicitly set "pm2" version using the "PM2_VERSION" environment
variable to ensure that an incompatible version is not accidentally
introduced in the future.

https://umd-dit.atlassian.net/browse/LIBCIR-290