LIBCIR-292. Added MD-SOAR favicon to "mdsoar" theme
Added MD-SOAR favicon to the "mdsoar" theme, which is a simple .ico
file.

In the "config/config.yaml" file, left commented out configuration for
various other variations of the favicon file, in the expectation that
we will eventually create these variants.

See https://evilmartians.com/chronicles/how-to-favicon-in-2021-six-files-that-fit-most-needs
for information about favicon variants.

https://umd-dit.atlassian.net/browse/LIBCIR-292