LIBCIR-294. Modified "config.server.ts" to read ".env" file
According to the DSpace 7 documentation at

https://wiki.lyrasis.org/display/DSDOC7x/User+Interface+Configuration

values in the ".env" file should be used as part of the environment
configuration, if such a file exists. This code change adds the
necessary initialization command as specified on the "dotenv"
package home page (https://github.com/motdotla/dotenv).

In DSpace 7.1, the "dotenv" initialization line was in
"scripts/set-env.ts", but that file doesn't exist in DSpace 7.2.

https://umd-dit.atlassian.net/browse/LIBCIR-294