LIBCIR-305. Added customized home page to "mdsoar" theme
Added “home-page” and “home-news” components to the “mdsoar” theme.

Heavily modified the
“src/themes/mdsoar/app/home-page/home-news/home-news.component.html”
file removing the background image, and adjusting the text to match that
on https://mdsoar.org. Also adjusted the “jumbotron” class in
“src/themes/mdsoar/app/home-page/home-news/home-news.component.scss” CSS
to provide better spacing.

Heavily modified the
“src/themes/mdsoar/app/home-page/home-page.component.html” file to more
closely resemble the home page on https://mdsoar.org/:

* Removed the “ds-top-level-community-list” displaying the list of communities
* Added a hard-coded “Browse” section, with links:
* Communities and Collections
* By Issue Date
* Author
* Titles
* Subjects
* Type

Added a “Submit” section, that enables a user to submit an item
(after logging in).

**Note:** Generating the “submit” page from this link seems very slow,
so it might be preferable to remove this section.

Added a hard-coded “Institutions in MD-SOAR” list, with links to each of
the MD-SOAR institutions, including
“University of Maryland, Eastern Shore” (see LIBCIR-282).

https://umd-dit.atlassian.net/browse/LIBCIR-305