LIBCIR-316. Adjusted Angular routing for "handle" redirects
On the home page, the hyperlinks to the individual institutional
communities in the “Institutions in MD-SOAR” list use the “handle” URLs
for the institutions (i.e., http://localhost:4000/handle/11603/1 for
Frostburg), instead of the “UUID” URLs
(i.e., http://localhost:4000/communities/e34f5843-e595-4050-9f20-3c730277abf8
for Frostburg).

DSpace properly redirects a “handle” URL to a “UUID” URL, but when using
the “Back” button in the browser, the user was not returned to the
original page.

Fixed this by changing the “findByIdAndIDType” method in the
“src/app/core/data/dso-redirect.service.ts” class, adding
"{ replaceUrl: true }" to the "router.navigate" call.

This suppresses the transient redirect from "handle" URL to "UUID" URL,
and ensures that left-click the "Back" button in the browser returns
the user to the original page.

Adjusted tests in "src/app/core/data/dso-redirect.service.spec.ts" to account for
"replaceUrl" NavigationExtra.

Added "docs/MdsoarAngularCustomizations.md" to record this change to
the stock DSpace Angular codebase.

https://umd-dit.atlassian.net/browse/LIBCIR-316