LIBCIR-336. Updated Simple Item sidebar with "Links to Files"
Added sidebar in the "simple item" page to show the "Links to Files"
field when "dc.description.uri" metadata is available.

This was a customization that had been made in DSpace 6 (see LIBCIR-78)
which was simply missed when doing the DSpace 7 customizations in
LIBCIR-322).

Updated the "en.json5" file with the "Links to Files" label text.

https://umd-dit.atlassian.net/browse/LIBCIR-336