LIBCIR-368. Roll back handle redirect customization from LIBCIR-316
In LIBCIR-316, the DSpace code was customized to redirect "handle" URLs
in such a way that the browser "Back" button properly returned to the
original page.

Code changes in DSpace 7.6 in the following commit/pull request:

* https://github.com/DSpace/dspace-angular/commit/d4a5308d0c7f845cbe0f0caa18437635e94a3067
* https://github.com/DSpace/dspace-angular/pull/2630

appear to make that customization no longer necessary (they also break
the current customization).

Restored the code changed in LIBCIR-316 to the DSpace 7.6.1 stock code.

https://umd-dit.atlassian.net/browse/LIBCIR-368