LIBCIR-372. Changed community theme identifiers from handles to UUIDs
Modified the identifiers used to associated a community with an
institutional theme from handle identifiers to UUIDs.

This is a workaround for the the stock DSpace bug described in
https://github.com/DSpace/dspace-angular/issues/2348
in which handles may not be properly parsed.

The stock DSpace bug is expected to be fixed in a 7.6.x update,
so simply commenting out the handles so that can be easily restored
when the bug is fixed.

Updated the "CommunityThemes.md" documentation to use UUID instead of
handle identifiers.

https://umd-dit.atlassian.net/browse/LIBCIR-372