LIBCIR-403. Back-port DSpace 8.0 fix for admin sidebar text cutoff
The following is an adaption of the DSpace 8.0 fix in the following
pull requests for use in DSpace 7.6.2:

* https://www.github.com/DSpace/dspace-angular/pull/2976
* https://www.github.com/DSpace/dspace-angular/pull/3004

The basic change is the same (detecting the browser and applying a
particular CSS change for Firefox), but the files that are modified
are different, due to changes in DSpace 8.0.

Also added `scrollbar-gutter: stable;` to the
"div#sidebar-top-level-items-container” selector in
“src/app/admin/admin-sidebar/admin-sidebar.component.scss”, to ensure
that the scrollbar width is always taken into account, even if the
scrollbar is not displayed, so that the menu item labels do not shift
based on the presence/absence of the scrollbar.

https://umd-dit.atlassian.net/browse/LIBCIR-403