LIBCIR-404. Enable text wrapping in navigation bar entries
This commit consists of the changes from dspace-angular Pull Request
3513, incorporated into the appropriate "mdsoar" and "dspace" theme
files.

The changes in Pull Request 3513 enable the entries in the navigation
bar to perform text wrapping at smaller browser widths (similar to the
behavior in DSpace 7.6.1) so that the right-most navigation bar entries
do not drop down to a second line and get obscured by the "hero" image.

https://umd-dit.atlassian.net/browse/LIBCIR-404