LIBCIR-423. Commented out "liveRegion" properties to pass unit tests
The unit tests were failing complaining that

```
Object literal may only specify known properties, and 'liveRegion' does not exist in type 'BuildConfig'.
```

Commented out the "liveRegion" properties (which don't have the related
component, and so it unused), so that the unit tests will pass.

This customization should be removed (and the "liveRegion" properties
restored) when migration to DSpace 7.6.3 or later.

https://umd-dit.atlassian.net/browse/LIBCIR-423