LIBCIR-450. Make "Subcommunities and Collections" default tab
Modified the "Browse" bar on the communities pages so that the
"Subcommunities and Collections" tab is selected by default,
instead of the "Search" tab.

This required modifying the routing, replacing the "subcoms-cols"
subpath that pointed to the "SubComColSectionComponent" with the
empty subpath, and adding a new subpath, "search" for pointing to the
"ComcolSearchSectionComponent".

Note: This would be a BREAKING CHANGE for anyone who had an MD-SOAR
search results page in their bookmarks – they would now get the
“Subcommunities and Collections” page (with the
“Subcommunities and Collections” tab highlighted), without any search
results.

https://umd-dit.atlassian.net/browse/LIBCIR-450